requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.1
reportlab>=4.0.4
pdfplumber>=0.9.0
docx2txt>=0.8
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pywin32>=306; sys_platform == "win32"
cairosvg>=2.5.0
Wand>=0.6.0
